Mobile automation ework/ework_automation/requirements.txt
pytest
pytest-bdd
appium-python-client
selenium
